requirements.txt
fastapi
uvicorn
groq
python-multipart
pillow
python-dotenv
streamlit
requests